Added cancel appointement and doctor availability

+++ template_java_project/src/com/healthmate/service/activity/CancelAppointmentActivity.java
@@ -5,12 +5,15 @@ import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.healthmate.service.dynamodb.AppointmentDao;
 import com.healthmate.service.dynamodb.UserDao;
 import com.healthmate.service.dynamodb.models.Appointment;
+import com.healthmate.service.dynamodb.models.Availability;
 import com.healthmate.service.dynamodb.models.User;
+import com.healthmate.service.exceptions.InvalidAttributeException;
 import com.healthmate.service.exceptions.UserNotFoundException;
 import com.healthmate.service.models.requests.CancelAppointmentRequest;
 import com.healthmate.service.models.response.CancelAppointmentResponse;
 
 import javax.inject.Inject;
+import javax.management.AttributeNotFoundException;
 import java.util.List;
 
 public class CancelAppointmentActivity implements RequestHandler<CancelAppointmentRequest, CancelAppointmentResponse> {
@@ -30,7 +33,10 @@ public class CancelAppointmentActivity implements RequestHandler<CancelAppointme
             throw new UserNotFoundException();
         }
         Appointment appointment = appointmentDao.getAppointment(cancelAppointmentRequest.getEmail(), cancelAppointmentRequest.getAppointmentTime());
-        getDoctorAvailability.updateAvailability(appointment);
+        if (appointment == null) {
+            throw new InvalidAttributeException("Appointment not exist");
+        }
+        getDoctorAvailability.updateAvailability(appointment, Availability.BOOKED, Availability.AVAILABLE);
         appointmentDao.deleteAppointment(appointment);
 
         return new CancelAppointmentResponse("Appointment got cancelled successfully");
